pyproject.toml
[project]
name = "flash-tta-research"
version = "1.0.0"
description = "FLASH-TT: Fast Layer-Analytic SHift-aware Test-Time adaptation Experiments"
authors = [{name = "AI Researcher"}]

dependencies = [
    "torch==2.1.0",
    "torchvision==0.16.0",
    "timm==0.9.12",
    "transformers==4.35.2",
    "datasets==2.15.0",
    "PyYAML==6.0.1",
    "scikit-learn==1.3.2",
    "pandas==2.1.4",
    "numpy==1.26.2",
    "pillow==10.1.0",
    "matplotlib==3.8.2",
    "seaborn==0.13.0",
    "fvcore==0.1.5.post20221221",
    "ptflops==0.7",
    "psutil==5.9.6",
    "pynvml==11.5.0"
]
